Fixed Mem::storew taking u8 instead of u16

+++ src/mem.rs
@@ -6,7 +6,7 @@ pub trait Mem {
     self.loadb(addr) as u16 | (self.loadb(addr + 1) as u16 << 8)
   }
 
-  fn storew(&mut self, addr: u16, val: u8) {
+  fn storew(&mut self, addr: u16, val: u16) {
     self.storeb(addr, (val & 0xff) as u8);
     self.storeb(addr + 1, ((val >> 8) & 0xff) as u8);
   }
